+++ karma.conf.js
@@ -36,7 +36,7 @@ module.exports = function (config) {
     reporters: ['progress', 'kjhtml'],
     port: 9876,
     colors: true,
-    logLevel: config.LOG_DEBUG,
+    logLevel: config.LOG_WARN,
     autoWatch: true,
     browsers: ['ChromeHeadless'],
     captureTimeout: 210000,