+++ src/app/app.routes.ts
@@ -30,6 +30,7 @@ import { CalendarDemoComponent } from './calendar/calendar.demo';
 import { CalendarLegendDemoComponent } from './calendar/calendar-legend.demo';
 import { CalendarUpdatedDemoComponent } from './calendar/calendar-updated.demo';
 import { CalendarRangeDemoComponent } from './calendar/calendar-range.demo';
+import { CardsActionableDemoComponent } from './cards/cards-actionable.demo';
 import { CardsDemoComponent } from './cards/cards.demo';
 import { CardsExpandableDemoComponent } from './cards/cards-expandable.demo';
 import { CardsMultiSelectDemoComponent } from './cards/cards-multi-select.demo';
@@ -255,6 +256,7 @@ export const routes: Routes = [
   { path: 'calendar-monthview-legend', component: CalendarLegendDemoComponent },
   { path: 'calendar-updated', component: CalendarUpdatedDemoComponent },
   { path: 'calendar-range', component: CalendarRangeDemoComponent },
+  { path: 'cards-actionable', component: CardsActionableDemoComponent },
   { path: 'cards', component: CardsDemoComponent },
   { path: 'cards-expandable', component: CardsExpandableDemoComponent },
   { path: 'cards-multi-select', component: CardsMultiSelectDemoComponent },