+++ src/app/about/about.demo.ts
@@ -18,7 +18,6 @@ export class AboutDemoComponent {
   }
 
   openAbout() {
-
     const VERSION_ATTR_NAME = 'data-ids-enterprise-ng-version';
     const version = $('html').attr(VERSION_ATTR_NAME);
     this.about = this.aboutService
@@ -31,6 +30,18 @@ export class AboutDemoComponent {
       .productName('Enterprise Angular Components')
       .version(`ver. ${version}`)
       .content('<p>Fashionable components for fashionable applications.</p>')
+      .beforeOpen((ref?: SohoAboutRef) => {
+        console.log('beforeopen', ref);
+        return true;
+      })
+      .closed((ref?: SohoAboutRef) => {
+        console.log('closed', ref);
+        return true;
+      })
+      .afterClose((ref?: SohoAboutRef) => {
+        console.log('afterclose', ref);
+        return true;
+      })
       .open();
   }