+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 13.1.0 Fixes
 
+- `[Datepicker]` Bug fix on initial range values in datepicker. ([#1200](https://github.com/infor-design/enterprise-ng/issues/1200))
 - `[Dropdown]` Updated readonly property to include readonly attr. ([#1189](https://github.com/infor-design/enterprise-ng/issues/1189))
 
 ### 13.1.0 Features